transfer geolocation into own component
